Always free SRAM allocation for spaceToDepth pass.
Previously SRAM allocated for spaceToDepth is not freed. As a
result, the compiler often fails to find a memory strategy for the
convolution layer after the spaceToDepth for the HW configurations with
small SRAM per EMC. The final output of spaceToDepth is always put in
DRAM and therefore the allocated SRAM is not needed.

To fix this, SRAM allocated for spaceToDepth is freed before
considering memory strategies for its succeding layers.

Signed-off-by: Ayan Halder <ayan.halder@arm.com>