Insert a format conversion layer after spaceToDepth
SpaceToDepth's output format is NHWC by design. It is not compatible
with a McePle pass using strategy X that requires its IFM in NHWCB. As a
result a network usually fails in compilation if it contains a
convolution layer taking IFM transformed by a large block size
spaceToDepth layer.

To fix this, a format conversion node converting NHWC to NHWCB is
inserted after the spaceToDepth.

Signed-off-by: Ayan Halder <ayan.halder@arm.com>