Replace MUL operator by DepthConv2D in the backend
ResNet V2_50 network has multiplication layers that are not at present
supported on the Ethos-N backend.

However, as ResNet is multiplying by a constant, a Depthwise Convolution
can be used to mimic the multiplication of tensors of shape NxHxWxC by
constants of form 1x1x1xC (1D tensor).

This patch implements a method for replacing unsupported layers that can
be expanded in the future to cover more cases.

Signed-off-by: Jakub Sujak <jakub.sujak@arm.com>