Ensure input stripe size is compatible with filter
If the input is not large enough for the filter it could result in
output being produced without doing a full convolution.

When selecting strategies and stripe sizes a check is added to make
sure that the input is large enough for the filter.

Without this restriction, wrong stripe size could be selected resulting
in output being produced without doing a full convolution.

Signed-off-by: Zingo Andersen <zingo.andersen@arm.com>