Dump command stream from driver library when requested
We can currently dump the Combined Memory Map (CMM) from the driver
library, but the developer is still required to manually run
an internal tool to convert the relevant parts to XML. This patch means
that internal developer builds of the driver library will include the
feature to dump the command stream directly to XML, saving time for
the developer.

For public builds this feature is not available and this patch has no
functional effect.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>