Add check to verify that weight stripes fit in tile
A check has been added to catch cases where the expected number of
weight stripes doesn't fit in the weight tile because the weight encoder
generated larger stripes than expected. The size of the weight tile is
decided using the worst case scenario for the weight encoding so this
check will indicate an issue in the encoder.

If the check fails, a new internal error exception will be thrown to
inform the user that an error has occurred. The Support Library's minor
version has been incremented to indicate the change in the API.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>