Fix RLE estimation in compression parameter finder
The weight compression parameter finder will always use RLE if there are
at least 20% zero weights. This causes issues in some cases because RLE
requires additional space to be encoded and if the weights can't be
compressed efficiently, the space needed for RLE can cause the total
size of the encoded weight data to go beyond the estimated worst case.

To resolve this, the finder will now always estimate both the case with
and without RLE when there are zero weights and use the most space
efficient one of the two. To ensure the correct parameters are selected
the estimation now accounts for the space needed for the RLE encoding
and the palette.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>