Fix GRC estimation to take padding into account
When the weight compression parameter finder estimates if it should use
uncompressed or GRC encoded weights, it does not take the padding
required to pack the GRC encoded weights into account. This causes the
finder in some cases to select GRC encoded weights when uncompressed
weights would be more efficient.

To fix this, the finder's GRC estimation has now been updated to take
the required padding into account.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>