Remove inference-time dependency on support library
Previously, the driver library had a dependency on the support library
as the user needed to provide a support_library::CompiledNetwork object
to it. This meant that the integration needed to load the support library
at inference-time, which is not desirable.

This review breaks that dependency by using a serialized binary format
to provide the compiled network to the driver library.

Public API changes:

The driver_library::Network constructor now takes a raw byte pointer which
it reads from to get the compiled network. It is no longer possible
to directly pass a support_library::CompiledNetwork.

The existing support_library::CompiledNetwork::Serialize() method is used
at network-compile time to serialize this binary stream. The format
of this binary stream has changed from the format previously used and so
existing serialized compiled networks are not compatible.

The support_library::DeserializeCompiledNetwork() function has been removed
as it should no longer be required and it would require additional effort to
keep this for backwards compatibility.

Implementation notes:

The (Compiled Network) deserialization code now lives solely in the
Driver Library - there is no longer any of this in the support library

The new binary format has less data than is required to fully represent
the support_library::CompiledNetwork, as the Driver Library does not need
all of the information that is available at network-compile time.

I have tried to implement this in a way that minimizes copying and
storing of the compiled network data, as this could be significant
for large networks on small platforms.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>