Check if reshape works as last layer in NHWCB format
Previously, test-cases have verified if reshape works
as last layer in NHWCB format with Startegy-3, i.e
small tensor of size (1, 16, 16, 16) was used as input
to these tests.

This patch adds an additional test-case to verify
that reshape works as last layer in NHWCB format when
the input tensors are large. The size of the input
tensors in the implemented test-case is (1, 256, 128, 16).
Larger tensor size forces Srategy-0 to be used instead
of Startegy-3.

Signed-off-by: Shubham <shubham@arm.com>