Fix Compilation for MaxPool 3x3 following deep convolution
In case of deep convolution followed by MaxPool 3x3 and the number of
input channels is too large the ifm will be split in width and since
the max pool PLE kernel does not support splitting in width the network
will fail to compile. This fix will insert identity depthwise before
the max pool whenever we find this pattern.

Signed-off-by: Hossam Ahmed <hossam.ahmed@arm.com>