Fix SMC call error checking
Error codes returned from the SMC calls are negatives values. However,
the value is stored as an unsigned value in the arm_smccc_res struct.
This is currently not taken into account when checking for errors from
the SMC calls.

To resolve this, the SMC call macros has been replaced with functions
that return the error code as a signed value. To further improve the
error checking, the NPU secure status is now bound checked before
returning the status and more descriptive error codes are now used.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>