Remove no longer needed support library CompiledNetwork public APIs
Now that the Compiled Network is passed to the Driver Library in the
serialized form, there is no need for the CompiledNetwork to have APIs
to query "internal" data such as the encoded weights.

These APIs have therefore been removed as they will be of no use to
users of the library and unnecessarily expose internal implementation
details.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>