Refactor ChooseAndSetupStrategy
A lot of parameter were passed to that function and forwarded to other
function. A struct StrategySelectionParameter has been introduced to
pack all of the parameters.

A StrategySelectionReturnValue structure has been created to pack the
the success flag, the SRAM allocator object and the SRAM tensor config.

TensorConfig has been renamed to StrategyConfig.

This is the first patch. A second patch targeting strategy X will
follow.
Signed-off-by: Laurent Carlier <laurent.carlier@arm.com>