Add reference counting to SRAM allocation.
A list of node id is maintained per allocated SRAM memory chunk. Each
can increase the reference counter by calling IncrementReferenceCount.
The reference count is automatically decremented when the chunk is freed
by a particular node.

In the case some nodes are grouped into one pass. I.e. they use the same
buffer but the intermediate node doesn't free the buffer, the last node
of the pass is used.

Signed-off-by: Laurent Carlier <laurent.carlier@arm.com>