Add a compatibility check between Driver Library and Kernel Module
The kernel module version is defined in ethosn.h.
The driver library checks for a compatible version of the kernel
module in the following scenarios:-

1. Compile time :- The driver library checks if the version of
the kernel module defined in ethosn.h is supported or not. The
supported versions are defined in KmodNetwork.hpp.

2. Run time :- The driver library checks if the version of the
kernel module defined in ethosn.h, matches with the one obtained
from the kernel.

To achieve this, driver library exposes an API VerifyKernel().
Users of the driver library (ie system-tests, armnn-ethos-n-backend)
are expected to invoke this API before they do any sort of interaction
with the kernel module.

Added tests to validate that VerifyKernel() returns True/False whenever
there is a match/mismatch.

Also removed unnecessary includes and bumped up the minor version of
the driver library (since a new API has been introduced).

Signed-off-by: Ayan Halder <ayan.halder@arm.com>