Fix an inference release bug in kernel module
Previously a core mutex is used to delete an inference from the list of
queued inferences. However an inference could be released before it is
completed and a "scheduled" inference may be in different states so that
it may or may not have acquired a core. As a result, the pointer to the
core could be a NULL pointer. In addition, it would have the risk of two
different threads trying to add and delete inferences from the list at
the same time, since inference queue mutex is used to add an inference
to the list.

To fix this issue, the inference queue mutex is used for deleting
inferences from the list of queued inferences.

Signed-off-by: Hui chen <hui.chen@arm.com>