Add version information to command stream
The command stream (header-only) library implicitly adds the current
version to a header before the commands. This current version is defined
by some #defines, which were already present.

The command stream library also reads this header when given a block
of memory to read as a command stream, and validates the version in the
data against the #defines. This means that the version number is
required to match exactly the one it was compiled for and there is no
backwards compatibility.

If we want to support reading other versions, more work is needed and
we'll need to figure out how to handle which parts happen in the command
stream library and which parts happen in the firmware.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>