Add a compatibility check between Support Library and Supported Command Stream
Ensures the version used by the support library is compatible
with the versions supported by a given set of firmware and
hardware capabilities.

Implemented as a function that compares the range given in a
set of capabilities with the version of the command stream
that the support library was compiled with.

Added tests to validate that VerifySupportedCommandStream() returns
true/false when the given capabilities are/aren't compatible with
the command stream version used.

Updated support libary access test in support library and backend
to compare against current version rather than updating
the set string used.

Signed-off-by: Joshua Pimm <joshua.pimm@arm.com>