Add libXmlToBinary in driver/support_library/command_stream/xml_to_binary
This contains the logic for converting command streams between binary
and XML formats.

Add scons option `allow_command_stream_dump` to enable the command
stream dumping feature (default OFF). This means that there will be no
additional build dependencies unless the user explicitly enables this
feature.

If the build option is enabled then the Driver Library will link against
libXmlToBinary and mxml.

Simplify the logic used to dump the command stream - it no longer
extracts it from the CMM but instead reads it directly from the
buffer in the CompiledNetwork.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>