Fix missing "strategy 1" plans
Some plans were not being correctly generated because of a bug with
limiting the number of stripes, that led to a negative-size range.

Add unit test to cover the fix and tweak the expected results for
mobilenet - it now cascades less layers, but by doing so achieves
a better DRAM bandwidth.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>