Weight encoder optimisation - replace map with array
The std::map requires heap allocations and tree traversal which is slow.
The new code uses a 256-element stack-allocated std::array which is
faster. To do this, we make the frequency table using the "raw" weights
(the uint8_t values that the user gives us) rather than the weight "symbols"
which are 16-bit values and would require much more space in the array.

Also use the same "raw" weights in FindRLEParams so that we don't need
to pass both "raw" and "uncompressed" weights to the functions. This
may also give some speedup as it's iterating over 8-bit elements rather
than 16-bit. Also attempt to optimize the calculation of zeroGroups
by removing the counter.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>