Fix scons default targets
With the move of libXmlToBinary into the support_library folder tree,
it was included in the scons default targets. This is unwanted as it
means that running scons without specifying targets will require
mxml to be installed which is an unnecessary dependency for most use
cases.

This patch changes the default targets to just the support_library itself
(and not any of it's sub-components) by using the build directory
as the default target, not the source directory.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>