Improve test categorisation
Move the mapping file test from EthosNReplacementTests.cpp to
EthosNMappingTests.cpp so that all the mapping file tests are together
and so that we don't have a mix of unrelated test cases in the same file.

This leaves EthosNReplacementTests.cpp with only test cases related to the
"replace unsupported" feature, for which the file has now been renamed.

I can now add new test cases to this file without adding to the confusion.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>