Fix core probing error checking in kernel module
During the parent device probing in the kernel module, it is assumed
that the call to populate the child devices (cores) will return an error
if the child device probe returns an error. This is not correct because
the return value from the child device probe is not checked inside
kernel's driver framework.

Because of this assumption, the character device will be created
regardless if the child device probing was successful or not, making it
look like the setup was successful when in fact there may not be any or
fewer cores than expected available.

To resolve this, the parent device probe will now verify that the
expected number of cores are available after the populate call. The
error handling in the parent device probe has also been changed to
ensure that the child devices are depopulated if an error occurs after
the populate call.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>