Move bias add replacement from support library to backend
The Support Library previously had a feature where it would claim
support for broadcasted addition, but this would only actually compile
if it was broadcasted addition with one input as a constant. Internally
the support library handled this by converting the addition to a
depthwise convolution.

In order to make the support library IsSupported checks more robust, the
support library no longer claims support for this case as it is not
possible to determine at IsSupported time if one of the inputs is
constant.

It is now expected that the integration is responsible for doing this
replacement, and in our case that means the Arm NN backend. Therefore
the Arm NN backend now detects this case and makes the appropriate
replacement.

The backend's IsAdditionSupported check now considers if the addition
can be supported via a depthwise replacement, if it can't be handled
natively by the support library.

If the addition layer is handed to the backend, we replace it with a
depthwise if that is appropriate.

This shares a lot of similarity with the existing Constant + Mul ->
Depthwise replacement which the backend does. Unfortunately this one is
slightly more complicated because we support addition natively as well,
whereas Mul is supported only via replacement.

Unit tests for all new functions have been added as well as a simple
end-to-end test in the backend.

The support library's IsAdditionSupported function no longer claims
support for any broadcasted addition, although I have left some of the
logic to make it easy to re-enable this later if necessary.

The support library's ReplaceConstantAdditionWithDepthwise function has
been removed as it no longer performs this optimisation.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>