Rework handling of global config in the backend
The existing code used some global variables to store global
configuration(e.g. the EthosNConfig) and these global variables were
both read and written in a variety of places and sometimes copied into
member variables, duplicating the information. Coupled with this, a lot
of the unit tests were creating config files and setting environment
variables to point to those files in order to configure the objects
under test.

All of this led to very confusing code and also meant that we were
reloading config files more often than we should be.

This patch reworks all this so that all of the objects that need config
information take it as parameters (e.g. in the constructor of the
object) rather than using global variables. This also means that tests
can configure the objects directly rather than resorting to setting
global state(i.e. we no longer need to set env vars or make files on
disk).

At the top-level, in the EthosNBackend object, we unfortunately still
need some global state due to the lack of an object with appropriate
lifetime to store the global config. However this is now contained in a
single location and is not exposed to the rest of the code (like the
previous globals were). There is also still some unfortunate
manipulation of this global state in the unit tests, which I don't
think can be avoided.

The new solution is not perfect but it is a lot easier to reason about
the new code as most of the smell is contained to a single place.

Implementation notes:

* Rename some functions which load files or make kernel calls to
make it clearer that they aren't cheap accessors:
    GetCapabilities() to QueryCapabilities()
    GetEthosNConfig() to ReadEthosNConfig()
    GetMappings() to ReadMappingsFromFile()
* Move EthosNBackendProfilingService::Instance() to cpp file to
avoid duplicate singleton across libraries

Signed-off-by: Rob Hughes <robert.hughes@arm.com>