Fix crash in IsAdditionSupported for mismatching data types
We have an assert in Addition::CalculateOutputTensorInfo which confirms
that the input data types are the same, but the IsSupported checks
were not checking this. Therefore if the user called the API with
mismatching data types, it would assert and crash.

This patch adds a proper check in IsAdditionSupported for this case
and adds a corresponding unit test.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>