Remove quotient limit from the RLE finder
Currently the RLE finder is limiting the number of quotient bits that
are allowed to be used. This is incorrect because the RLE compression
does not have such a limit and can cause the parameter finder to opt to
not use RLE when the number of zeros in a group is very large.

To resolve this, the limit has now been removed.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>