Fix crash in backend unit test
This unit test for the layer mapping feature is designed to check that a
mapping file which mentions a pooling layer but does not specify the
function (i.e. max vs avg) raises an appropriate error. However this
test was crashing before it got to the point of testing this, because
it failed to create the input graph on which it was going to perform
the replacement. This is because it tried to create a pooling layer
without a specified function.

To fix this I have simply swapped the mapping file for this test so that
the pooling layer is the *replacement* rather than the original graph.
This means it does not attempt to create the invalid pooling layer
(it creates an activation layer instead). The pooling layer is then
validated as expected when it tries to do the replacement.

We are potentially losing some coverage for validation of the pattern
(as opposed to the replacement), but I can't see any evidence that this
was intended here (all the other tests in the same block of code test
validation of the replacement, not the pattern). Either way, this test
was incorrect and needs fixing.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>