Support Constant operation as regular input
Previously, Constant operations were usable as the weights/bias for
convolution-like operations and would fail to compile if used as the
'regular' input for any operation.

This patch fixes this and so they are fully supported at any point in
the graph. This is a simple case of placing the constant data in
the "ConstantDmaData" section of the compiled network (like encoded
weights) - the rest of the driver stack already handles everything
needed.

Update unit tests.

No changes to the Arm NN backend are required as it already supports
Constant layers.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>