Fix a crash in the Ethos-N Backend while using mapping
In armnn::ParseMappings(), one should always check if the buf is empty
or not. The reason being if the buffer is empty, it should not create
any entry in mappingsFromFile.
Furthermore, an exception is thrown to the user if the mapping file does
not conform to the following format.

pattern:
blah blah
graph-replacement:
tah tah

The user can denote comments in the mapping file as a line starting with "#".
The parser ignores the comments.
If the mapping file contains only comments or is empty, an appropriate warning
is thrown.

Signed-off-by: Ayan Halder <ayan.halder@arm.com>