Reduce stream overhead in weight encoder
Currently when all the streams for an OG stripe is collected the
streams' data is copied rather than moved and the stream padding is
performed in two steps. First the streams are padded to have equal
length and then they are padded again to be DMA aligned.

To reduce some of this overhead, the streams' data is now moved instead
of copied and the padding is performed in one step.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>