Add bit width based GRC/RLE divisor selection
When the weight compression parameter finder estimates what GRC and RLE
divisor to use, it always tests all possible divisors regardless of the
values. This is unnecessary because depending on the range of values
that should to be divided, many of the divisors can be ruled out right
away as they will always give a sub optimal or unsupported result.

The finder will now use the bit width of the smallest and largest value
that should be divided to calculate what divisors that are worth
testing. To be able to do this check in a more efficent way, the finder
has been changed to keep track of the smallest and largest values for
GRC and RLE.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>