Support None/Medium/High for DUMP_DEBUG_FILES in config file
The support library's m_DumpDebugFiles option was previously upgraded
from a bool to an enum, but the Arm NN backend config file was left
as a boolean, mapping to None/High. There was no way to select
Medium through the backend.

This patch updates the backend config file to support values of
None/Medium/High, as well as the existing values of 0/1 to preserve
backwards compatibility.

It also removes the unused operator<< rather than updating it.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>