Improve debugging for failed cascading estimation
Previously, the EstimateOpGraph would throw an exception if any part
of the OpGraph could not be estimated and there was no simple way
of determining which part of the OpGraph was causing the problem.

It now returns information even if it fails, by making the return object
also capable of representing partially estimated OpGraphs. An IsComplete()
method can be used on the returned object to check if the estimation
was fully succcessful or not.

This means we can dump the partially estimated OpGraph to dot files, and
this shows exactly which parts of the graph failed to estimate.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>