Add and use GetPerformanceMetrics()
The Combiner was previously hardcoding a list of performance metrics
for debugging purposes, but these could easily become out-of-sync with
the metrics used for actual performance comparison. It now uses a new function
which uses a common list of metrics so it will be easier to keep in sync.

This is in preparation for tweaking the metrics in a following patch.

Also refactor IsLeftMoreDataPerformantThanRight() to be more flexible.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>