Replace Passes metric with MceCycleCount
Cascading will now prefer Combinations that have a lower total MCE cycle
count (if the DRAM bandwidth is the same). This is an improved
replacement over the existing "number of Passes" metric because it will
account for layers using Winograd vs Direct convolution(Winograd should
result in less cycle count).

Signed-off-by: Rob Hughes <robert.hughes@arm.com>