Change Dram ReinterpretNode plans to no-ops
This enables reshape support in some cases (reshape as last layer),
and moves towards full reshape support in the future.

Rather than the original plan of using DummyOp, we now treat Dram
reinterprets as true no-ops, by not having any Ops in the Plan at all.
This is a better representation because Ops are intended to represent
runtime operations, and reinterpreting a Dram buffer is not a runtime
operation at all. As a concrete example of why this is better, it
removes the need to "estimate" these DummyOps. The new plan
is just a single Dram buffer, tagged as *both* input and output, so that
when combined with adjacent plans, there are no new Ops added to the
OpGraph.

This change alters the way we treat some properties of
Buffers as now the m_TensorShape is a bit less rigid and does not need
to exactly match between adjacent plans. This relaxation allows the
Combiner to find compatible plans with the above representation.

Going forward, the idea is that the buffer properties describe
*one possible way* to interpret the data (perhaps a canonical
representation?), but Ops that use the Buffer are free to reinterpret
those properties as appropriate. In this case it means that a Dram
NHWC buffer's shape can be interpreted to be something different,
as long as it has the same number of elements. The same concept can
be used in the future for reinterpeting quantization, for example.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>