Replace MulByScalar with ReinterpretQuantization
Uptil now, support for multiplication with a constant was
provided only for the cases where constant is of shape
(1, 1, 1, C1), input is of shape (N, H, W, C2) and C1 == C2.

This implementation adds supports for the additional case where
C1 != C2 and C1 == 1. This support is provided by replacing
multiplication operation in Backend with a ReinterpretQuantization
operation from the Support Library.

Additionally, unit tests have also been implemented.

Signed-off-by: Shubham <shubham@arm.com>