Enforce NHWCB for concat output.
The firmware does not support DMA-ing NHWC when splitting in depth. The
output of the concat operation will be forced to be NHWCB by adding
FormatConversionNode at the output of the concat if it is NHWC (similar
to what we are doing in input nodes). This change will enable split
in depth for both MCE and PLE operations following a concat.

Signed-off-by: Hossam Ahmed <hossam.ahmed@arm.com>