Document and validate carveout properties
When using a carveout (i.e. no SMMU), the user-defined memory carveout
region must meet certain criteria regarding alignment and size. If
these criteria are not met then the driver stack will not function
correctly.

This patch adds documentation and validation for these criteria, so that
the kernel module will gracefully error out if they are not met. The
criteria regarding speculative reads cannot be verified by the kernel
module so there is no validation or graceful failure for this.

The kernel module also now always prints a debug message showing the
location and size of the carveout, to aid with debugging.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>