Re-queue ongoing inference during power suspension
The state of the inference needs to be consistent with the inference
queue content. If this is not the case things can go wrong.
The approach used in the suspend method puts the system in an
inconsistent state if the inference is aborted before going into RUNNING
state after resume.
Use inference queue instead.

Signed-off-by: Gabriele Vitiello <gabriele.vitiello@arm.com>