Move to ad-hoc plan generation
Changed CreatePlans() to GetPlans() in part. This now returns the plans
instead of storing them inside the part. This enables plans to be
created when they are needed. Also it doesn't throw if there are 0
plans generated.

This doesn't change the number or types of plans generated. Just where
they are returned and stored.

Plans are now a shared_ptrs instead of unique_ptrs because the
Combinations need to reference them.

Currently this doesn't cache any plans.

Change visualization and dumping to use plans directly instead of parts

Caching and changing the interface still needs to be done.

Signed-off-by: Joshua Slater <joshua.slater@arm.com>