Fix uninitialized tensor shape use without asserts
The tensor shape for the weight stripe is initialized according to the
weight's data format and if the format is unknown an assert will be
triggered. However, if asserts are disabled, the code will continue to
use the uninitialized tensor shape.

To handle this regardless if asserts are enabled or not, an exception
will now be thrown as well.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>