Change Cascading part interface
Change Part to be a base class which the old Part class (now PartV1) inherits.

GraphOfParts now controls the layout of parts instead of parts having
knowledge about connected parts.

Don't use edges/nodes in parts and the
combiner.

Change the visualisation to use parts instead of edges and nodes but
still display the nodes in the parts for debuggability.

Add tests for the graph of parts.

Signed-off-by: Joshua Slater <joshua.slater@arm.com>