Use IsConstant flag in backend
Arm NN TensorInfos have an "IsConstant" flag which was not previously
being used in the backend (it was added relatively recently). This
patch makes use of this flag by checking that it is set for tensors
which we require to be constant (weights and biases). If it is not set
in these cases, then the layer is reported as not supported.

This is a prerequisite for advertising the ConstantTensorsAsInputs
capability, as this will expose the backend to non-constant
weights/bias tensors which we will therefore need to handle correctly
(by rejecting them).

Signed-off-by: Rob Hughes <robert.hughes@arm.com>