Support for latest Arm NN
Arm NN have removed some backend APIs so we need to remove our
implementation in order to be compatible with newer Arm NN versions. We
already implement the new equivalent APIs, so there is no corresponding
new code to add.

Add implementation of IsChannelShuffleSupported and
IsConvolution3dSupported, required for implementing the ILayerSupport
interface.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>