Fix overflow in MCE direct conv cycle estimation
When estimating the number of cycles needed to complete a MCE direct
convolution, several 32-bit unsigned integers are multiplied and the
result is stored in a 32-bit unsigned integer which may overflow.

To resolve this, the result will now be stored in a 64-bit unsigned
integer and a 64-bit multiplication is performed instead.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>