Fix core address not set before SMC call
In the non-secure variant of the kernel module, the SMC call to get the
secure status of an core currently fails because the core's address is
not setup until after the SMC call so an invalid address is used in the
call.

To address this issue, the SMC function has been changed to take the
address to the core separately so it can be used before setting up the
core. The reset SMC function has been changed in a similar manner to
keep the functions consistent.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>