Add McePart and FusedPlePart
McePart and FusedPlePart inherits the BasePart class for Mce and Fused
Ple Operations.

Implements GetPlans with beginning, middle, end and lonely cascades.

Add a StripeHelper file with stripe generator which is used to generate
stripes for the beginning and lonely plans. This is used in McePart and
FusedPlePart. It generates the compute and memory stripe shapes.

Remove m_Location from the DmaOp class. Dma Ops should be between two
buffers which have locations, so having the location stored in the Op
is unnecessary.

Change PartUtils to be free standing function instead of a class with
static methods, and to be in the impl namespace.

Signed-off-by: Joshua Slater <joshua.slater@arm.com>