Fix DMA allocator being destroyed twice
If something fails after a DMA allocator has been created during the
device probing, the allocator will be cleaned up. However, the pointer
to the allocator is not cleaned up so later when the device is removed,
the remove function will attempt to cleaned up the allocator again and
access memory that has already been freed.

To resolve this, the pointer to the allocator will now be set to NULL
when the allocator is cleaned up.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>