Add check for max number of profiling hw counters
Currently the number of hardware counters in the profiling config is not
checked before the config is sent to the firmware. To ensure that the
number of counters specified can't be larger than what the config can
hold, the kernel module will now check the number of counters before
sending the config to the firmware.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>