Remove shared_ptr from Part::GetPlans interface
Previously, Part::GetPlans returned an array of shared_ptrs to plans,
which was because the Plans were being stored in a cache. We no longer
have the cache, so there is no reason to return shared_ptrs. This makes
the interface simpler, as can be seen by the removal of various pointer
dereferences and ".get()" calls in this patch.

The Combiner does however need to move the Plans into shared_ptrs,
because there can be more than one Combination that refers to the same
Plan, and we want to avoid duplicating these quite large objects.

Also delete old Combiner.hpp

Signed-off-by: Rob Hughes <robert.hughes@arm.com>