Add NetworkToGraphOfPartsConverter for Cascading
This patch implements the new conversion step to go directly
from a Network to a Graph of Parts.

An algorithm was implemented to walk the Operations in the Network,
generate the appropriate new Part types and finally append them to
a Graph of Parts. Only the basic Operations necessary for VGG are
supported at the moment such as Input, Output, Convolution,
Pooling (MAX) and Reshape. Additional Operations will be addded in
future patches.

A new Unit-Test was added in order to check the functionality of
the NetworkToGraphOfPartsConverter. The test creates a simple Network
and converts it to a GraphOfParts using the new Converter. It then
checks whether the GraphOfParts was constructed and connected
correctly.

Signed-off-by: Konstantinos Karavelas <konstantinos.karavelas@arm.com>