Remove calls to virtual method in constructor
The Node class was making a call to its virtual method Reset in its
constructor, which has unexpected behaviour and so is not good
practice.

This was being done to make sure that the first iteration of preparation
was done with all the Nodes in the same "reset" state as in all later
iterations (Reset was called *after* each iteration of preparation, so
the first iteration was special).

This patch removes the call in the constructor and instead relies on the
client of the Node class (the Compiler class) to call Reset before the
first iteration.

Member fields are now initialized in the class initializer lists, as
they were previously being set by Reset(), which is no longer being
called in the constructor.

Add missing call to base class Reset in InputNode's implementation

Rename Reset to ResetPreparation as a more descriptive name.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>