Make exception types nothrow copy-constructible
Several custom exception types were using std::string, which does not
provide a noexcept copy-constructor, and therefore the copy-constructor
of the custom exception type was itself not nothrow. This is bad
practice as an exception during copying of an exception object will
lead to the program being terminated.

This patch changes the custom exception types to inherit from
std::runtime_error, which stores a message and provides a noexcept
copy-constructor. This also simplifies our code because we can use that
class as-is.

Bump patch version of Support Library as this is a backwards-compatible
bug fix. Driver Library version has already been bumped for the next
release.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>