Add Visitor function for Concatenation in NetworkToGraphOfPartsConverter
This patch adds support for Concatenation Operations in the new
NetworkToGraphOfPartsConverter.

A new Visitor function was implemented to handle Concatenation
Operations in a Network. More specifically, the Visitor generates
the appropriate Part types and appends them to the Graph Of Parts.
For each input in the ConcatPart, if that input's quantization
information is different from the output's, an McePart is added
to do the requantization.

Signed-off-by: Konstantinos Karavelas <konstantinos.karavelas@arm.com>